requirements.txt
Appium-Python-Client~=5.1.1
pytest-bdd~=8.1.0
selenium~=4.34.2
pytest~=8.4.1